inline game for referencegame; adjust referencegame to new player behavior
